Various upgrades to logging etc.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/base/motorcontrol/MotionProfile.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.base.motorcontrol;
 
 public interface MotionProfile {
-    abstract double runProfile(double time);
+    abstract int    runProfile(double time);
     abstract void   resetProfile(double Amax_in,
                                  double Vmax_in,
                                  double Vi_in,
